Code refactor - command pattern
