+++ .vscode/launch.json
@@ -14,7 +14,6 @@
             "cwd": "${workspaceFolder}/server-side",
             "program": "${workspaceFolder}/server-side/app.local.ts",
             "outFiles": [
-                "${workspaceFolder}/server-side/build/*/*.js"
             ],
             "preLaunchTask": {
                 "label": "Build API",
