+++ server-side/PfsCommands/AtomicCommands/downloadFileCommand.ts
@@ -1,0 +1,9 @@
+import AbstractCommand from '../abstractCommand';
+
+export class downloadFileCommand extends AbstractCommand 
+{
+
+	public async execute(): Promise<any>{
+		return await this.downloadFile();
+	}
+}
