+++ server-side/PfsCommands/AtomicCommands/listObjectsCommand.ts
@@ -1,0 +1,15 @@
+import AbstractCommand from '../abstractCommand';
+
+export class ListObjectsCommand extends AbstractCommand 
+{
+
+	public async execute(): Promise<any>{
+		return await this.listObjects();
+	}
+
+	private async listObjects(whereClause?:string)
+	{
+		//TODO order by creation date, to support DIMX export.
+		return this.pfsGetter.getObjects(whereClause);
+	}
+}
